+++ src/router/elegant/transform.ts
@@ -97,10 +97,13 @@ function transformElegantRouteToVueRoute(
     if (component) {
       if (isSingleLevelRoute(route)) {
         const { layout, view } = getSingleLevelRouteComponent(component);
-  
+
         const singleLevelRoute: RouteRecordRaw = {
           path,
           component: layouts[layout],
+          meta: {
+            title: route.meta?.title || ''
+          },
           children: [
             {
               name,
@@ -110,36 +113,35 @@ function transformElegantRouteToVueRoute(
             } as RouteRecordRaw
           ]
         };
-  
+
         return [singleLevelRoute];
       }
-  
+
       if (isLayout(component)) {
         const layoutName = getLayoutName(component);
-  
+
         vueRoute.component = layouts[layoutName];
       }
-  
+
       if (isView(component)) {
         const viewName = getViewName(component);
-  
+
         vueRoute.component = views[viewName];
       }
-  
+
     }
   } catch (error: any) {
     console.error(`Error transforming route "${route.name}": ${error.toString()}`);
     return [];
   }
 
-  
   // add redirect to child
   if (children?.length && !vueRoute.redirect) {
     vueRoute.redirect = {
       name: children[0].name
     };
   }
-  
+
   if (children?.length) {
     const childRoutes = children.flatMap(child => transformElegantRouteToVueRoute(child, layouts, views));
 
@@ -210,6 +212,7 @@ const routeMap: RouteMap = {
   "plugin_editor_markdown": "/plugin/editor/markdown",
   "plugin_editor_quill": "/plugin/editor/quill",
   "plugin_excel": "/plugin/excel",
+  "plugin_gantt": "/plugin/gantt",
   "plugin_icon": "/plugin/icon",
   "plugin_map": "/plugin/map",
   "plugin_pdf": "/plugin/pdf",